requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Migrate==4.1.0
Flask-Bootstrap==3.3.7.1
Flask-Admin==1.6.1
python-dotenv==1.1.1
WTForms==3.2.1
SQLAlchemy==2.0.43
Werkzeug==3.1.3
alembic==1.16.5